feat: update app structure
