+++ package.json
@@ -50,7 +50,7 @@
     "koa-json": "^2.0.2",
     "koa-router": "^10.1.1",
     "koa2-swagger-ui": "^5.3.0",
-    "swagger-ui-express": "4.1.6",
+    "swagger-ui-express": "^4.2.0",
     "uuid": "^8.3.2",
     "yamljs": "0.3.0"
   },
