+++ src/resources/boards/board.router.ts
@@ -1,12 +1,12 @@
 import Router from 'koa-router';
-import { handlers } from './board.handler';
+import * as services from './board.service';
 
 export const router = new Router();
 
 router
-  .get('/boards', handlers.getAllBoards)
-  .get('/boards/:boardId', handlers.getBoard)
-  .post('/boards', handlers.createBoard)
-  .put('/boards/:boardId', handlers.updateBoard)
-  .delete('/boards/:boardId', handlers.deleteBoard);
+  .get('/boards', services.getAllBoards)
+  .get('/boards/:boardId', services.getBoard)
+  .post('/boards', services.createBoard)
+  .put('/boards/:boardId', services.updateBoard)
+  .delete('/boards/:boardId', services.deleteBoard);
