+++ src/resources/db/board.db.ts
@@ -1,0 +1,3 @@
+import { Board } from '../types/types';
+
+export const boardRepo: Map<string, Board> = new Map();
