+++ src/resources/db/task.db.ts
@@ -1,0 +1,3 @@
+import { Task } from '../types/types';
+
+export const taskRepo: Map<string, Map<string, Task>> = new Map();
