+++ src/resources/db/user.db.ts
@@ -1,0 +1,3 @@
+import { User } from '../types/types';
+
+export const usersRepo: Map<string, User> = new Map();
