+++ src/resources/users/user.router.ts
@@ -1,11 +1,11 @@
 import Router from 'koa-router';
-import { handlers } from './user.handler';
+import * as services from './user.service';
 
 export const router = new Router();
 
 router
-  .get('/users', handlers.getAllUsers)
-  .get('/users/:userId', handlers.getUser)
-  .post('/users', handlers.createUser)
-  .put('/users/:userId', handlers.updateUser)
-  .delete('/users/:userId', handlers.deleteUser);
+  .get('/users', services.getAllUsers)
+  .get('/users/:userId', services.getUser)
+  .post('/users', services.createUser)
+  .put('/users/:userId', services.updateUser)
+  .delete('/users/:userId', services.deleteUser);
